Map actions to wheel speeds
Removes debug prints and returns raw actions, instead mapping velocity commands to wheel angular speeds.
Adds a helper function to compute left/right wheel speeds via robot kinematics and adjusts the environment step accordingly.

+++ fast_td3/train_dclp_isaac_lab.py
@@ -192,7 +192,7 @@ def main():
                         device_type=amp_device_type, dtype=amp_dtype, enabled=amp_enabled
                     ):
                         actions = dclp.get_action(norm_obs)
-                    obs, reward, done, info = envs.step(actions.float())
+                    obs, reward, done, info = envs.step(actions)
                     episode_return += reward.mean().item()
                     episode_length += 1
                     # Update step progress bar with current reward
